Game of Life simulation in javascript
This is a simple Game of Life simulation written in javascript as a
study for how the simulation works.  It features two versions and two
types of each.  The first version is pretty basic and the second adds
more features.  For each version there is a zip containing jquery and a
non zipped version which uses a jquery CDN (this version requires being
connected to the internet).